Split devtools shell
